apply all comments
