+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/controller/AssetController.java
@@ -9,7 +9,6 @@ import org.orange.oie.internship2025.assetmanagementsystem.dto.*;
 import org.orange.oie.internship2025.assetmanagementsystem.entity.AssetCategory;
 import org.orange.oie.internship2025.assetmanagementsystem.entity.AssetType;
 import org.orange.oie.internship2025.assetmanagementsystem.service.serviceInterface.*;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
@@ -24,16 +23,14 @@ import java.util.List;
 public class AssetController {
 
     private final AssetAssignmentService assetAssignmentService;
-    private AssetDetailsService assetDetailsService;
     private final AssetService assetService;
     private final AssetTypeService typeService;
     private final CategoryService categoryService;
 
-    public AssetController(AssetService assetService, AssetTypeService typeService, CategoryService categoryService, AssetDetailsService assetDetailsService, AssetAssignmentService assetAssignmentService) {
+    public AssetController(AssetService assetService, AssetTypeService typeService, CategoryService categoryService, AssetAssignmentService assetAssignmentService) {
         this.assetService = assetService;
         this.typeService = typeService;
         this.categoryService = categoryService;
-        this.assetDetailsService = assetDetailsService;
         this.assetAssignmentService = assetAssignmentService;
     }
 
@@ -86,7 +83,16 @@ public class AssetController {
 
     @GetMapping("/details/{id}")
     public ResponseEntity<AssetDetailsDto> getAssetDetails(@PathVariable Long id) {
-        return ResponseEntity.ok(assetDetailsService.getAssetDetails(id));
+        return ResponseEntity.ok(assetService.getAssetDetails(id));
+    }
+
+    @GetMapping("/history/{assetId}")
+    public ResponseEntity<Page<listAssetHistoryResponseDto>> getAssetHistoryByAssetId(
+            @PathVariable Long assetId,
+            Pageable pageable
+    ) {
+        Page<listAssetHistoryResponseDto> historyPage = assetService.getHistoryByAssetId(assetId, pageable);
+        return ResponseEntity.ok(historyPage);
     }
 
 }
