+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/dto/listAssetHistoryResponseDto.java
@@ -1,16 +1,13 @@
 package org.orange.oie.internship2025.assetmanagementsystem.dto;
 
-import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
-import org.orange.oie.internship2025.assetmanagementsystem.entity.Asset;
-import org.orange.oie.internship2025.assetmanagementsystem.entity.User;
 import org.orange.oie.internship2025.assetmanagementsystem.enums.AssetStatus;
 
 import java.time.LocalDateTime;
 @Setter
 @Getter
-public class AssetHistoryDto {
+public class listAssetHistoryResponseDto {
     private Long id;
     private String assetName;
     private String assignedTo;
