+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/mapper/AssetMapper.java
@@ -1,13 +1,7 @@
 package org.orange.oie.internship2025.assetmanagementsystem.mapper;
 
-import org.orange.oie.internship2025.assetmanagementsystem.dto.AssetDto;
-import org.orange.oie.internship2025.assetmanagementsystem.dto.AssetRequestDto;
-import org.orange.oie.internship2025.assetmanagementsystem.dto.UpdateAssetDto;
-import org.orange.oie.internship2025.assetmanagementsystem.dto.UserDTO;
-import org.orange.oie.internship2025.assetmanagementsystem.entity.Asset;
-import org.orange.oie.internship2025.assetmanagementsystem.entity.AssetCategory;
-import org.orange.oie.internship2025.assetmanagementsystem.entity.AssetType;
-import org.orange.oie.internship2025.assetmanagementsystem.entity.User;
+import org.orange.oie.internship2025.assetmanagementsystem.dto.*;
+import org.orange.oie.internship2025.assetmanagementsystem.entity.*;
 import org.orange.oie.internship2025.assetmanagementsystem.repository.CategoryRepository;
 import org.orange.oie.internship2025.assetmanagementsystem.repository.TypeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -108,4 +102,47 @@ public class AssetMapper {
         }
            return dtoList;
     }
+
+    public AssetDetailsDto toDetailsDto(Asset asset) {
+        AssetDetailsDto dto = new AssetDetailsDto();
+        dto.setAssetId(asset.getId());
+        dto.setAssetName(asset.getName());
+        dto.setBrand(asset.getBrand());
+        dto.setAssetDescription(asset.getDescription());
+        dto.setCategoryName(asset.getCategory().getName());
+        dto.setTypeName(asset.getType().getName());
+        dto.setLocation(asset.getLocation());
+        dto.setSerialNumber(asset.getSerialNumber());
+        dto.setPurchaseDate(asset.getPurchaseDate().toString());
+        dto.setWarrantyEndDate(asset.getWarrantyEndDate().toString());
+        dto.setStatus(asset.getStatus());
+        dto.setImagePath(asset.getImagePath());
+
+        if (asset.getAssignments() != null && !asset.getAssignments().isEmpty()) {
+            var userAssignment = asset.getAssignments().get(asset.getAssignments().size() - 1);
+            User assignedUser = userAssignment.getAssignedTo();
+            dto.setAssignedToId(assignedUser.getId());
+            dto.setAssignedToName(assignedUser.getUsername());
+            dto.setAssignedToEmail(assignedUser.getEmail());
+        }
+
+        return dto;
+    }
+
+    public listAssetHistoryResponseDto toHistoryDto(AssetHistory history) {
+        listAssetHistoryResponseDto dto = new listAssetHistoryResponseDto();
+        dto.setId(history.getId());
+        dto.setAssetName(history.getAsset().getName());
+        dto.setNote(history.getNote());
+        dto.setTimestamp(history.getTimestamp());
+        dto.setStatus(history.getStatus());
+        if (history.getUser() != null) {
+            dto.setAssignedTo(history.getUser().getUsername());
+        } else {
+            dto.setAssignedTo(null);
+        }
+
+        return dto;
+    }
+
 }
