+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/service/serviceImpl/AssetServiceImpl.java
@@ -132,4 +132,19 @@ public class AssetServiceImpl implements AssetService {
         List<Asset> availableAssets = assetRepository.findAll(spec);
         return assetMapper.toDtoList(availableAssets);
     }
+
+    @Override
+    public AssetDetailsDto getAssetDetails(Long id) {
+        Asset asset = assetRepository.findById(id)
+                .orElseThrow(() -> new BusinessException(ApiReturnCode.ASSET_NOT_FOUND,
+                        "Asset not found with id " + id));
+
+        return assetMapper.toDetailsDto(asset);
+    }
+
+    @Override
+    public Page<listAssetHistoryResponseDto> getHistoryByAssetId(Long assetId, Pageable pageable) {
+        Page<AssetHistory> histories = assetHistoryRepository.findByAssetId(assetId, pageable);
+        return histories.map(assetMapper::toHistoryDto);
+    }
 }
