+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/service/serviceInterface/AssetDetailsService.java
@@ -1,8 +1,8 @@
-package org.orange.oie.internship2025.assetmanagementsystem.service.serviceInterface;
-
-import org.orange.oie.internship2025.assetmanagementsystem.dto.AssetDetailsDto;
-
-
-public interface AssetDetailsService {
-    AssetDetailsDto getAssetDetails(Long id);
-}
+//package org.orange.oie.internship2025.assetmanagementsystem.service.serviceInterface;
+//
+//import org.orange.oie.internship2025.assetmanagementsystem.dto.AssetDetailsDto;
+//
+//
+//public interface AssetDetailsService {
+//    AssetDetailsDto getAssetDetails(Long id);
+//}
