+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/service/serviceInterface/AssetHistoryService.java
@@ -1,8 +1,8 @@
-package org.orange.oie.internship2025.assetmanagementsystem.service.serviceInterface;
-import org.orange.oie.internship2025.assetmanagementsystem.dto.AssetHistoryDto;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-public interface AssetHistoryService {
-    Page<AssetHistoryDto> getHistoryByAssetId(Long assetId, Pageable pageable);
-}
+//package org.orange.oie.internship2025.assetmanagementsystem.service.serviceInterface;
+//import org.orange.oie.internship2025.assetmanagementsystem.dto.listAssetHistoryResponseDto;
+//import org.springframework.data.domain.Page;
+//import org.springframework.data.domain.Pageable;
+//
+//public interface AssetHistoryService {
+//    Page<listAssetHistoryResponseDto> getHistoryByAssetId(Long assetId, Pageable pageable);
+//}
