+++ src/main/java/org/orange/oie/internship2025/assetmanagementsystem/service/serviceInterface/AssetService.java
@@ -12,4 +12,6 @@ public interface AssetService {
     List<AssetDto> getAllAssets();
     List<AssetDto> getAvailableAsset(String type);
     Page<ListAssetDTO> getFilteredAsset(AssignedAssetFilterDTO filterDTO, Pageable pageable);
+    AssetDetailsDto getAssetDetails(Long id);
+    Page<listAssetHistoryResponseDto> getHistoryByAssetId(Long assetId, Pageable pageable);
 }
